refactor: moves feed functions into feed/ directory
