+++ src/lib/feed/save-feed-item-for-deleted-event.js
